completed test fixes for resource and time refactors
