+++ kivakit-filesystems/s3fs/src/test/java/com/telenav/kivakit/filesystems/s3fs/S3FileSystemServiceTest.java
@@ -19,7 +19,7 @@
 package com.telenav.kivakit.filesystems.s3fs;
 
 import com.telenav.kivakit.core.progress.ProgressReporter;
-import com.telenav.kivakit.core.test.support.SlowTest;
+import com.telenav.kivakit.test.SlowTest;
 import com.telenav.kivakit.test.UnitTest;
 import com.telenav.kivakit.resource.FileName;
 import org.junit.Before;
