+++ kivakit-filesystems/s3fs/src/test/java/com/telenav/kivakit/filesystems/s3fs/S3FolderTest.java
@@ -18,7 +18,7 @@
 
 package com.telenav.kivakit.filesystems.s3fs;
 
-import com.telenav.kivakit.core.test.support.SlowTest;
+import com.telenav.kivakit.test.SlowTest;
 import com.telenav.kivakit.test.UnitTest;
 import com.telenav.kivakit.filesystem.Folder;
 import com.telenav.kivakit.filesystem.spi.FolderService;
